move to Util namespace
